+++ config/test.exs
@@ -1,5 +1,15 @@
 use Mix.Config
 
 config :faktory_worker_ex,
-  client_config: ClientConfig,
-  worker_config: WorkerConfig
+  port: 7421,
+  client: [pool: 5],
+  workers: [
+    default: [
+      concurrency: 2
+    ],
+    middleware: [
+      concurrency: 2,
+      middleware: [BadMath],
+      queues: ["middleware"]
+    ]
+  ]