+++ lib/faktory/configuration/client.ex
@@ -1,0 +1,9 @@
+defmodule Faktory.Configuration.Client do
+  @moduledoc false
+
+  defstruct [
+    host: "localhost", port: 7419, pool: 10, middleware: [], fn: nil,
+    name: "default", wid: nil
+  ]
+
+end