+++ lib/faktory/supervisor.ex
@@ -7,17 +7,18 @@ defmodule Faktory.Supervisor do
   end
 
   def init(nil) do
+    alias Faktory.Configuration
 
     # Only start the workers from the mix task.
     children = if Faktory.start_workers? do
-      [{Faktory.Supervisor.Workers, Faktory.worker_config_module}]
+      [{Faktory.Supervisor.Workers, Configuration.fetch(:worker)}]
     else
       []
     end
 
     # Always start the clients supervisor.
     children = [
-      {Faktory.Supervisor.Clients, Faktory.client_config_module}
+      {Faktory.Supervisor.Clients, Configuration.fetch(:client)}
       | children
     ]