+++ test/support/client_config.ex
@@ -1,6 +1,0 @@
-defmodule ClientConfig do
-  use Faktory.Configuration, :client
-
-  port 7421
-  pool 5
-end