+++ test/test_helper.exs
@@ -1,6 +1,7 @@
 Faktory.flush
 {:ok, _} = PidMap.start_link
 {:ok, _} = Stack.start_link
-{:ok, _} = Faktory.Supervisor.Workers.start_link(WorkerConfig)
-{:ok, _} = Faktory.Supervisor.Workers.start_link(WorkerConfigMiddleware)
+Faktory.Configuration.fetch_all(:worker) |> Enum.each(fn config ->
+  {:ok, _} = Faktory.Supervisor.Workers.start_link(config)
+end)
 ExUnit.start()